Do chores
- Code restructure.
- Remove service interfaces to avoid overhead complexity.
- Fix JavaDocs.
