--- src/main/java/com/ittovative/otpservice/service/SmsService.java
@@ -1,13 +1,0 @@
-package com.ittovative.otpservice.service;
-
-import com.ittovative.otpservice.dto.OtpRequestDto;
-
-public interface SmsService {
-    /**
-     * Send string.
-     *
-     * @param otpRequestDto the otp request dto
-     * @return the string
-     */
-    String send(OtpRequestDto otpRequestDto);
-}
